% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Application}
\alias{Application}
\title{Simulation}
\usage{
Application(seed, iterations = 55000, burnin = 5000, thinning = 50,
  L = 35, K = 55)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
35 simulations
}
\examples{
Simulation(seed=1)

}

