% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Applications}
\alias{Applications}
\title{Simulation}
\usage{
Applications(seed, iterations = 55000, burnin = 5000, thinning = 20,
  L = 35, K = 20)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
35 simulations
}
\examples{
Simulation(seed=1)

}

