% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\docType{class}
\name{DP-class}
\alias{DP-class}
\title{Dirichlet Process
DP applied to censored data}
\description{
Dirichlet Process
DP applied to censored data
}
\examples{
\dontrun{
weights <- matrix(c(0,1,0,0,1,0,0,1,0), ncol=3)
data <- sim.data(weights)
 G1 <- init.DP(G1, prior=list(mu=0, n=0.1, v=3, vs2=1*3), L=35, thinning=2,
 burnin = 0, max_iter = 5000, clustering = T )
 G1 <- MCMC.DP(G1, data, 500)
validate.DP(G1, data)
}

}

