\name{DPsurv-package}
\alias{DPsurv-package}
\alias{DPsurv}
\docType{package}
\title{
\packageTitle{DPsurv}
}
\description{
\packageDescription{DPsurv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DPsurv}
\packageIndices{DPsurv}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DPsurv}

Maintainer: \packageMaintainer{DPsurv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
plot(rnorm(5))
}
