% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDP.R
\docType{class}
\name{HDP-class}
\alias{HDP-class}
\title{Hierarchical Dirichlet Process
HDP applied to censored data}
\description{
Hierarchical Dirichlet Process
HDP applied to censored data
}
\examples{
\dontrun{
weights <- matrix(c(1,0,0,0,1,0,0,0,1), ncol=3)
data <- sim.data(weights)
G3 <- new("HDP")
G3 <- init.HDP(G3, prior=list(mu=0, n=0.1, v=3, vs2=1*3), L=15, 
                   J=length(unique(data@presentation$Sample)), thinning=2,
                   burnin = 5000, max_iter = 55000)
G3 <- MCMC.HDP(G3, data, 55000)
plot.ICDF(G3@theta, G3@phi, G3@weights[,1], G3@L, grid=0:500,
           distribution=data@presentation, xlim=500)
validate.HDP(G3, data)
}

}

