% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncnorm}
\alias{rtruncnorm}
\title{Random Number From a Truncated Normal}
\usage{
rtruncnorm(y, mu, sigma, K = 1L)
}
\arguments{
\item{y}{is the value at which we want to left-truncated our normal distribution}

\item{mu}{is the mean of the truncated normal}

\item{sigma}{is the s.d. of the truncated normal}

\item{K}{is the number of replication of draws}
}
\description{
Draw from a truncated normal
}

