% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Application}
\alias{Application}
\title{Application}
\usage{
Application(seed, iterations = 35000, burnin = 5000, thinning = 50,
  L = 55, K = 35)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
Produce the experiment on the performance art dataset.
}
\examples{
\dontrun{
Application(seed=1)
}

}

