% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\docType{class}
\name{DP-class}
\alias{DP-class}
\title{Dirichlet Process
DP applied to censored data}
\description{
Dirichlet Process
DP applied to censored data
}
\examples{
\dontrun{
weights <- matrix(c(1,0,0,0,1,0,0,0,1), ncol=3)

data <- sim.data(n=100, J=10, weights)

G1 <- new("DP")

G1 <- init.DP(G1, DataStorage=data, prior=list(mu=0, n=0.1, v=3, vs2=1*3), L=35, thinning=50,
          burnin = 5000, max_iter = 55000 )
          
G1 <- MCMC.DP(G1, data, 55000)

validate.DP(G1, data)
 
plot.ICDF(G1@theta, G1@phi, G1@weights, G1@L, grid=0:500,
           distribution=data@presentation, xlim=500)
}

}

