% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDP.R
\docType{class}
\name{NDP-class}
\alias{NDP-class}
\title{Nested Dirichlet Process
NDP applied to censored data}
\description{
Nested Dirichlet Process
NDP applied to censored data
}
\examples{
\dontrun{
weights <- matrix(c(1,0,0,0,1,0,0,0,1), ncol=3)

data <- sim.data(n=100, J=10, weights)

G2 <- init.NDP(prior=list(mu=0, n=0.1, v=3, vs2=1*3),K=5, L=35, thinning=50,
              burnin = 0, max_iter = 5000 )
              
G2 <- MCMC.NDP(G2, data, 500)

validate.NDP(G2, data)

plot.ICDF(G2@theta[,which.max(G2@pi)], G2@phi[,which.max(G2@pi)], G2@weights[,which.max(G2@pi)],
            G2@L, grid=0:500, distribution=data@presentation, xlim=500)
}

}

