% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Simulation}
\usage{
Simulation(seed, replications = 35, iterations = 55000, burnin = 5000,
  thinning = 50, L = 35, K = 20, n = 100, J = 10, case = "1",
  factor = 1)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
35 simulations
}
\examples{
Simulation(seed=1)

}

