% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Illustration of crayon colors}
\usage{
Simulation(seed, replications = 35, iterations = 55000, burnin = 5000,
  thinning = 20, L = 35, K = 20, n = 100, J = 10, case = 1,
  factor = 1)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
Simulation(seed=1)

}

