% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Simulation}
\usage{
Simulation(seed, replications = 32, iterations = 55000, burnin = 5000,
  thinning = 50, L = 35, K = 20, n = 100, J = 10, case = "1",
  factor = 1, Prior = c(0, 0.01, 2 * 3, 2 * 1, rep(c(5, 0.1), 2)),
  val_prop = 0, nb_cores = 4)
}
\arguments{
\item{seed}{}
}
\value{
None
}
\description{
35 simulations
}
\examples{
Simulation(seed=1)

}

